Set up test assets once for the whole test suite rather than for each test
