+++ full-stack-tests/presentation/scripts/setup-tests.js
@@ -4,13 +4,16 @@
  *--------------------------------------------------------------------------------------------*/
 
 const chai = require("chai");
+const cpx = require("cpx2");
 const faker = require("faker");
+const fs = require("fs");
 const chaiJestSnapshot = require("chai-jest-snapshot");
 const chaiAsPromised = require("chai-as-promised");
 const sinonChai = require("sinon-chai");
 const chaiSubset = require("chai-subset");
 const jsdom = require("jsdom");
 const sourceMapSupport = require("source-map-support");
+const path = require("path");
 
 console.log(`Backend PID: ${process.pid}`);
 
@@ -30,6 +33,40 @@ chai.use(sinonChai);
 chai.use(chaiSubset);
 
 faker.seed(1);
+
+// Set up assets
+cpx.copySync(`assets/**/*`, "lib/assets");
+cpx.copySync(`public/**/*`, "lib/public");
+copyITwinBackendAssets("lib/assets");
+copyITwinFrontendAssets("lib/public");
+function copyITwinBackendAssets(outputDir) {
+  const iTwinPackagesPath = "node_modules/@itwin";
+  fs.readdirSync(iTwinPackagesPath)
+    .map((packageName) => {
+      const packagePath = path.resolve(iTwinPackagesPath, packageName);
+      return path.join(packagePath, "lib", "cjs", "assets");
+    })
+    .filter((assetsPath) => {
+      return fs.existsSync(assetsPath);
+    })
+    .forEach((src) => {
+      cpx.copySync(`${src}/**/*`, outputDir);
+    });
+}
+function copyITwinFrontendAssets(outputDir) {
+  const iTwinPackagesPath = "node_modules/@itwin";
+  fs.readdirSync(iTwinPackagesPath)
+    .map((packageName) => {
+      const packagePath = path.resolve(iTwinPackagesPath, packageName);
+      return path.join(packagePath, "lib", "public");
+    })
+    .filter((assetsPath) => {
+      return fs.existsSync(assetsPath);
+    })
+    .forEach((src) => {
+      cpx.copySync(`${src}/**/*`, outputDir);
+    });
+}
 
 beforeEach(function () {
   const currentTest = this.currentTest;
